tests/wpt/web-platform-tests/html/webappapis/dynamic-markup-insertion/document-write/module-delayed.html

<!doctype html>
<title>async document.write in a module</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  // Expose {test} in the iframe for using the step_timeout helper.
  document.test = t;
  const iframe = document.createElement("iframe");

  iframe.onerror = t.unreached_func("Error loading iframe");
  let onLoadWasCalled = false;
  iframe.onload = t.step_func(() => {
    assert_equals(iframe.contentDocument.body.textContent, "Initial body contents\n");
    // Don't call the event handler another time after document.write.
    iframe.onload = null;
    onLoadWasCalled = true;
  });
  document.addEventListener("documentWriteDone", t.step_func_done(() => {
    assert_equals(iframe.onload, null);
    assert_equals(iframe.contentDocument.body.textContent, "Initial body contents\n");
  }));

  iframe.src = "module-delayed-iframe.html";
  document.body.appendChild(iframe);
});
</script>
