src/third_party/blink/web_tests/external/wpt/html/webappapis/dynamic-markup-insertion/document-write/module-delayed-expected.txt

This is a testharness.js-based test.
FAIL async document.write in a module assert_equals: expected "Initial body contents\n" but got "document.write body contents\n"
Harness: the test ran to completion.
